Adjust directory structure
We now separate out the node/electron elements of the app from the
python elements. This will make it easier to keep concerns separated
and work on one aspect of the application at a time. Someday, if we
decide to separate the interface and the backend into separate
repositories, they will already be prepared.