Reenable VTT file generation in our whisper backend
This renders this branch functional! The full app now works in dev
and uses our custom python backend.

Ultimately we will want to disable VTT file generation and do
communication between electron and the backend using either stdio
or a shared database, but this stopgap gets us a functional app